requirements.txt
requests
pyobjc
pyobjc-framework-SystemConfiguration